Renamed e2e to acceptance in AdminX
