+++ apps/admin-x-settings/test/e2e/advanced/integrations/amp.test.ts
@@ -1,5 +1,5 @@
 import {expect, test} from '@playwright/test';
-import {globalDataRequests, mockApi, updatedSettingsResponse} from '../../../utils/e2e';
+import {globalDataRequests, mockApi, updatedSettingsResponse} from '../../../utils/acceptance';
 
 test.describe('AMP integration', async () => {
     test('Supports toggling and filling in AMP integration', async ({page}) => {
