+++ apps/admin-x-settings/test/e2e/advanced/integrations/custom.test.ts
@@ -1,6 +1,6 @@
 import {Integration, IntegrationsResponseType} from '../../../../src/api/integrations';
 import {Webhook, WebhooksResponseType} from '../../../../src/api/webhooks';
-import {chooseOptionInSelect, globalDataRequests, limitRequests, mockApi, responseFixtures} from '../../../utils/e2e';
+import {chooseOptionInSelect, globalDataRequests, limitRequests, mockApi, responseFixtures} from '../../../utils/acceptance';
 import {expect, test} from '@playwright/test';
 
 test.describe('Custom integrations', async () => {
