+++ apps/admin-x-settings/test/e2e/general/publicationLanguage.test.ts
@@ -1,5 +1,5 @@
 import {expect, test} from '@playwright/test';
-import {globalDataRequests, mockApi, updatedSettingsResponse} from '../../utils/e2e';
+import {globalDataRequests, mockApi, updatedSettingsResponse} from '../../utils/acceptance';
 
 test.describe('Publication language settings', async () => {
     test('Supports editing the locale', async ({page}) => {
